+++ lib/regexp_parser/expression/classes/backref.rb
@@ -1,9 +1,12 @@
 module Regexp::Expression
   module Backreference
-    class Base < Regexp::Expression::Base; end
+    class Base < Regexp::Expression::Base
+      attr_accessor :referenced_expression
+    end
 
     class Number < Backreference::Base
       attr_reader :number
+      alias reference number
 
       def initialize(token, options = {})
         @number = token.text[token.token.equal?(:number) ? 1..-1 : 3..-2].to_i
@@ -13,6 +16,7 @@ module Regexp::Expression
 
     class Name < Backreference::Base
       attr_reader :name
+      alias reference name
 
       def initialize(token, options = {})
         @name = token.text[3..-2]
@@ -20,27 +24,31 @@ module Regexp::Expression
       end
     end
 
-    class NumberCall         < Backreference::Number; end
-    class NumberRelative     < Backreference::Number; end
-    class NumberCallRelative < Backreference::Number; end
-    class NameCall < Backreference::Name; end
+    class NumberRelative     < Backreference::Number
+      attr_accessor :effective_number
+      alias reference effective_number
+    end
 
-    class NumberRecursionLevel < Backreference::Base
-      attr_reader :number, :recursion_level
+    class NumberCall         < Backreference::Number; end
+    class NameCall           < Backreference::Name; end
+    class NumberCallRelative < Backreference::NumberRelative; end
+
+    class NumberRecursionLevel < Backreference::Number
+      attr_reader :recursion_level
 
       def initialize(token, options = {})
-        @number, @recursion_level = token.text[3..-2].split(/(?=[+-])/).map(&:to_i)
         super
+        @number, @recursion_level = token.text[3..-2].split(/(?=[+-])/).map(&:to_i)
       end
     end
 
-    class NameRecursionLevel < Backreference::Base
-      attr_reader :name, :recursion_level
+    class NameRecursionLevel < Backreference::Name
+      attr_reader :recursion_level
 
       def initialize(token, options = {})
+        super
         @name, recursion_level = token.text[3..-2].split(/(?=[+-])/)
         @recursion_level = recursion_level.to_i
-        super
       end
     end
   end