+++ lib/regexp_parser/expression/classes/conditional.rb
@@ -7,6 +7,8 @@ module Regexp::Expression
     end
 
     class Condition < Regexp::Expression::Base
+      attr_accessor :referenced_expression
+
       # Name or number of the referenced capturing group that determines state.
       # Returns a String if reference is by name, Integer if by number.
       def reference
@@ -18,6 +20,8 @@ module Regexp::Expression
     class Branch < Regexp::Expression::Sequence; end
 
     class Expression < Regexp::Expression::Subexpression
+      attr_accessor :referenced_expression
+
       def <<(exp)
         expressions.last << exp
       end